Crate bmw_evh

The BMW Eventhandler crate

This crate defines and implements the crate::EventHandler trait. The event handler handles events on tcp/ip connections. It manages both inbound and outbound connections. The underlying mechanism used are Epoll on Linux, Kqueues on MacOS and WePoll on Windows. These libraries allow for perfromant handling of reads and writes on multiple socket connections.

Motivation

Eventhandler provides an fast interface to the low level eventing libraries on various platforms. It will be the basis for the HTTP server and the rustlet library and eventually the cryptocurrency that will be built on top of these libraries.

Performance

The /etc directory in the projects inlcudes a subdirectory called “evh_perf”. This subdirectory is used for testing the performance of the eventhandler.

Examples

Macros

evh

The crate::evh! macro builds a crate::EventHandler implementation. It is returned as a Box<dyn EventHandler<OnRead, OnAccept, OnClose, OnHousekeeper, OnPanic> + Send + Sync>. If an error occurs, a bmw_err::Error is rerurned.

evh_oro

The crate::evh_oro! macro builds a crate::EventHandler implementation. This is a simplified version of the crate::evh! macro. It allows the user to not have to specify the handlers other than the crate::EventHandler::set_on_read handler. Hence, it is called evh_oro (evh on read only). This macro is useful for testing, but also for simple use cases where only on read processing is necessary. It can allow the user to avoid specifying the unused boilerplate code that is required since all of the handlers must be specified with the crate::evh! macro. The result is returned as a Box<dyn EventHandler<OnRead, OnAccept, OnClose, OnHousekeeper, OnPanic> + Send + Sync>. If an error occurs, a bmw_err::Error is rerurned.

Structs

Chunk

A chunk of data returned by the crate::EventHandler. Chunks are of a maximum size defined by the EvhReadSlabSize parameter minus 4 bytes.

Connection

The crate::Connection struct represents a connection. It may be either a server side connection or a client side connection. To create a server side connection, see crate::EvhBuilder::build_server_connection. To create a client side connection, see crate::EvhBuilder::build_client_connection. These connections can then be added to a crate::EventHandler via the crate::EventHandler::add_server_connection and crate::EventHandler::add_client_connection functions respectively.

EvhBuilder

Builder struct for the crate. All implementations are created through this struct.

EvhStats

Statistical information for the crate::EventHandler. This struct may be retrieved by calling the crate::EventHandler::wait_for_stats function.

WriteHandle

The crate::WriteHandle struct may be used to write data to the underlying connection. Since std::clone::Clone is implmeneted, the WriteHandle may be cloned and passed to other threads. This allows for asynchronous writes for applications like Websockets.

Traits

EventHandler

The crate::EventHandler trait is implemented by the returned value of the crate::EvhBuilder::build_evh function.

UserContext

The crate::UserContext trait is returned on all callbacks specified by the crate::EventHandler. It can be used to read data from the underlying connections, clear slabs that are not needed, and get or set a user_data structure that can be used as a context variable by the caller. The user_data structure is of the type Box<dyn Any + Send + Sync> so the user may use this for practically anything. This value stays consistent accross all callbacks for the thread that is invoked on. Each thread of the crate::EventHandler does store and return a distinct value. See the crate documentation as well for the background information and motivation for this crate as well as examples. See crate::EventHandler for the callback functions.