Macro bmw_evh::evh_oro

macro_rules! evh_oro {
    ($($config:tt)*) => { ... };
}

The crate::evh_oro! macro builds a crate::EventHandler implementation. This is a simplified version of the crate::evh! macro. It allows the user to not have to specify the handlers other than the crate::EventHandler::set_on_read handler. Hence, it is called evh_oro (evh on read only). This macro is useful for testing, but also for simple use cases where only on read processing is necessary. It can allow the user to avoid specifying the unused boilerplate code that is required since all of the handlers must be specified with the crate::evh! macro. The result is returned as a Box<dyn EventHandler<OnRead, OnAccept, OnClose, OnHousekeeper, OnPanic> + Send + Sync>. If an error occurs, a [bmw_err::Error] is rerurned.

Input Parameters

EvhThreads (usize) (optional) - The number of threads for this crate::EventHandler. The default value is 4.

EvhReadSlabSize (usize) (optional) - The size of the slabs, in bytes, in the read slab allocator. The read slab allocator is where request data is temporarily stored before it can be processed and cleared. Each thread has it’s own dedicated slab allocator. The default value is 512.

EvhReadSlabCount (usize) (optional) - The count of the slabs, in the read slab allocator. The read slab allocator is where request data is temporarily stored before it can be processed and cleared. Each thread has it’s own dedicated slab allocator. The default value is 1_000.

EvhTimeout (u16) (optional) - The time, in milliseconds that the event handler will wait to get events, if none occur. This value is important for things like the house keeping frequency and stats configurations because if no events occur and this value is too great, a house keeper or stats action will not occur until this timeout occurs. If too small a number is used, too many needless events occur, if too big, the periodic tasks will not occur until later than desired. The default value is set to 1_000 (1 second).

EvhHouseKeeperFrequencyMillis (usize) (optional) - The frequency, in milliseconds, at which the housekeeper callback is called. The housekeeper closure is set by the crate::EventHandler::set_on_housekeeper function. It is important to note that each thread executes it’s own housekeeper function. The default value is 10_000 (10 seconds).

EvhStatsUpdateMillis (usize) - The frequency, in milliseconds, at which the stats data is returned. The stats data may be retrieved by calling the crate::EventHandler::wait_for_stats function. The default value is 5_000 (5 seconds).

Debug (bool) - If this parameter is set to true, additional debugging information will be logged. This parameter must NOT be set in a production configuration.

Returns

A Ok(Box<dyn EventHandler<OnRead, OnAccept, OnClose, OnHousekeeper, OnPanic> + Send + Sync>) is returned on success. A [bmw_err::Error] is returned on failure.

Errors

[bmw_err::ErrKind::Configuration] - If any values are specified other than the allowed values mentioned above or if there are any duplicate parameters specified.

[bmw_err::ErrKind::Configuration] - If EvhReadSlabCount is 0.

[bmw_err::ErrKind::Configuration] - If EvhReadSlabSize is less than 25.

[bmw_err::ErrKind::Configuration] - If EvhTimeout is 0.

[bmw_err::ErrKind::Configuration] - If EvhHouseKeeperFrequencyMillis is 0.

See also

See the crate documentation as well for the background information and motivation for this crate as well as examples.

Examples

 use bmw_err::*;
 use bmw_test::*;
 use bmw_evh::*;
 use bmw_log::*;
 use std::str::from_utf8;

 info!();

 fn main() -> Result<(), Error> {
     // create an evh with the specified configuration. The unspecified values are
     // the defaults mentioned above.
     let mut evh = evh_oro!(
         EvhTimeout(100),
         EvhThreads(1),
         EvhReadSlabSize(100),
         EvhStatsUpdateMillis(5000),
         Debug(true)
     )?;

     // set the on read handler
     evh.set_on_read(move |connection, ctx| -> Result<(), Error> {
         // loop through each of the available chunks and append data to a vec.
         let mut data: Vec<u8> = vec![];

         loop {         
             let next_chunk = ctx.next_chunk(connection)?;
             cbreak!(next_chunk.is_none());
             let next_chunk = next_chunk.unwrap();
             data.extend(next_chunk.data());
         }  

         // convert returned data to a utf8 string
         let dstring = from_utf8(&data)?;
         info!("data[{}]='{}'", connection.id(), dstring)?;

         // get a write handle
         let mut wh = connection.write_handle()?;

         // echo
         wh.write(dstring.as_bytes())?;

         // clear all chunks from this connection. Note that partial
         // clearing is possible with the ctx.clear_through function
         // or no data can be cleared at all in which case it can
         // be accessed on a subsequent request. When the connection
         // is closed, all data is cleared automatically.
         ctx.clear_all(connection)?;

         Ok(())
     })?;

     // no other handlers are necessary

     evh.start()?;

     Ok(())
 }