Trait bmw_evh::ConnData

pub trait ConnData {
    fn tid(&self) -> usize;
    fn get_connection_id(&self) -> u128;
    fn get_handle(&self) -> RawFd;
    fn get_accept_handle(&self) -> Option<RawFd>;
    fn write_handle(&self) -> WriteHandle;
    fn borrow_slab_allocator<F, T>(&self, f: F) -> Result<T, Error>
    where
        F: FnMut(&Box<dyn SlabAllocator + Send + Sync>) -> Result<T, Error>;
    fn slab_offset(&self) -> u16;
    fn first_slab(&self) -> u32;
    fn last_slab(&self) -> u32;
    fn clear_through(&mut self, slab_id: u32) -> Result<(), Error>;
}

This trait which is implemented by crate::ConnectionData. This trait is used to interact with a connection.

Required Methods

fn tid(&self) -> usize

Returns the thread id that this event occurred on.

fn get_connection_id(&self) -> u128

Returns the connection id of this event.

fn get_handle(&self) -> RawFd

Returns the handle of this event.

fn get_accept_handle(&self) -> Option<RawFd>

Returns the handle (if any) which this connection was accepted on.

fn write_handle(&self) -> WriteHandle

Returns a write handle which can be used to write to this connection.

fn borrow_slab_allocator<F, T>(&self, f: F) -> Result<T, Error> where
F: FnMut(&Box<dyn SlabAllocator + Send + Sync>) -> Result<T, Error>,

borrows, immutably, the slab allocator associated with this connection.

fn slab_offset(&self) -> u16

Returns the offset that data has been read to in the most recent slab for this connection.

fn first_slab(&self) -> u32

Returns the first slab in which data has been read for this connection.

fn last_slab(&self) -> u32

Returns the last slab in which data has been read for this connection.

fn clear_through(&mut self, slab_id: u32) -> Result<(), Error>

Clears data in slabs from memory up to and including slab_id for this connection.