Trait bmw_evh::EventHandler

pub trait EventHandler<OnRead, OnAccept, OnClose, HouseKeeper, OnPanic> where
    OnRead: FnMut(&mut ConnectionData<'_>, &mut ThreadContext) -> Result<(), Error> + Send + 'static + Clone + Sync + Unpin,
    OnAccept: FnMut(&mut ConnectionData<'_>, &mut ThreadContext) -> Result<(), Error> + Send + 'static + Clone + Sync + Unpin,
    OnClose: FnMut(&mut ConnectionData<'_>, &mut ThreadContext) -> Result<(), Error> + Send + 'static + Clone + Sync + Unpin,
    HouseKeeper: FnMut(&mut ThreadContext) -> Result<(), Error> + Send + 'static + Clone + Sync + Unpin,
    OnPanic: FnMut(&mut ThreadContext, Box<dyn Any + Send>) -> Result<(), Error> + Send + 'static + Clone + Sync + Unpin, {
    fn set_on_read(&mut self, on_read: OnRead) -> Result<(), Error>;
    fn set_on_accept(&mut self, on_accept: OnAccept) -> Result<(), Error>;
    fn set_on_close(&mut self, on_close: OnClose) -> Result<(), Error>;
    fn set_housekeeper(&mut self, housekeeper: HouseKeeper) -> Result<(), Error>;
    fn set_on_panic(&mut self, on_panic: OnPanic) -> Result<(), Error>;
    fn stop(&mut self) -> Result<(), Error>;
    fn start(&mut self) -> Result<(), Error>;
    fn add_client(
        &mut self, 
        connection: ClientConnection
    ) -> Result<WriteHandle, Error>;
    fn add_server(&mut self, connection: ServerConnection) -> Result<(), Error>;
}

This trait defines the behaviour of an eventhandler. See the module level documentation for examples.

Required Methods

fn set_on_read(&mut self, on_read: OnRead) -> Result<(), Error>

sets the on read callback for this crate::EventHandler. This callback is exuected whenever data is read by a connection.

fn set_on_accept(&mut self, on_accept: OnAccept) -> Result<(), Error>

sets the on accept callback for this crate::EventHandler. This callback is executed whenever a connection is accepted.

fn set_on_close(&mut self, on_close: OnClose) -> Result<(), Error>

sets the on close callback for this crate::EventHandler. This callback is executed whenever a connection is closed.

fn set_housekeeper(&mut self, housekeeper: HouseKeeper) -> Result<(), Error>

sets the housekeeper callback for this crate::EventHandler. This callback is executed once every crate::EventHandlerConfig::housekeeping_frequency_millis milliseconds. Each thread executes this callback independantly.

fn set_on_panic(&mut self, on_panic: OnPanic) -> Result<(), Error>

sets the on panic callback for this crate::EventHandler. This callback is executed whenever a user callback generates a thread panic.

fn stop(&mut self) -> Result<(), Error>

Stop this crate::EventHandler and free all resources associated with it.

fn start(&mut self) -> Result<(), Error>

Start the crate::EventHandler.

fn add_client(
&mut self,
connection: ClientConnection
) -> Result<WriteHandle, Error>

Add a crate::ClientConnection to this crate::EventHandler.

fn add_server(&mut self, connection: ServerConnection) -> Result<(), Error>

Add a crate::ServerConnection to this crate::EventHandler.