Trait bmw_evh::UserContext

pub trait UserContext {
    // Required methods
    fn next_chunk(
        &mut self,
        connection: &mut Connection,
    ) -> Result<Option<Chunk<'_>>, Error>;
    fn clear_all(&mut self, connection: &mut Connection) -> Result<(), Error>;
    fn clear_through(
        &mut self,
        slab_id: usize,
        connection: &mut Connection,
    ) -> Result<(), Error>;
    fn get_user_data(&mut self) -> &mut Option<Box<dyn Any + Send + Sync>>;
    fn set_user_data(&mut self, user_data: Box<dyn Any + Send + Sync>);
}

The crate::UserContext trait is returned on all callbacks specified by the crate::EventHandler. It can be used to read data from the underlying connections, clear slabs that are not needed, and get or set a user_data structure that can be used as a context variable by the caller. The user_data structure is of the type Box<dyn Any + Send + Sync> so the user may use this for practically anything. This value stays consistent accross all callbacks for the thread that is invoked on. Each thread of the crate::EventHandler does store and return a distinct value. See the crate documentation as well for the background information and motivation for this crate as well as examples. See crate::EventHandler for the callback functions.

Required Methods

fn next_chunk( &mut self, connection: &mut Connection, ) -> Result<Option<Chunk<'_>>, Error>

Get the next chunk of data, if available, that has been read by the crate::EventHandler for this crate::Connection.

Input Parameters

connection - the crate::Connection to get the next chunk from.

Returns

On success, the next crate::Chunk or None if there are no more chunks. [bmw_err::Error] is returned if an error occurs.

See Also

crate, crate::UserContext, crate::UserContext::clear_all

fn clear_all(&mut self, connection: &mut Connection) -> Result<(), Error>

Clear all slabs that are associated with this crate::Connection.

Input Parameters

connection - The crate::Connection to clear data from.

Returns

On success, unit is returned and on failure, [bmw_err::Error] is returned.

See Also

crate, crate::UserContext, crate::UserContext::clear_through, crate::UserContext::next_chunk

fn clear_through( &mut self, slab_id: usize, connection: &mut Connection, ) -> Result<(), Error>

Clear all slabs, through the slab_id specified, that are associated with this crate::Connection.

Input Parameters

slab_id - The slab_id of the slab to clear data through. This value can be obtained by calling crate::UserContext::next_chunk and then calling the crate::Chunk::slab_id function. connection - The crate::Connection to clear data from.

Returns

On success, unit is returned and on failure, [bmw_err::Error] is returned.

See Also

crate, crate::UserContext, crate::UserContext::clear_all, crate::UserContext::next_chunk

fn get_user_data(&mut self) -> &mut Option<Box<dyn Any + Send + Sync>>

Get the user_data object associated with this thread.

Input Parameters

none

Returns

The user_data which has been specified on a previous call to crate::UserContext::set_user_data.

See Also

crate, crate::UserContext, crate::UserContext::set_user_data

fn set_user_data(&mut self, user_data: Box<dyn Any + Send + Sync>)

Set the user_data object associated with this thread.

Input Parameters

user_data - The value to set for this thread’s context data.

Returns

none

See Also

crate, crate::UserContext, crate::UserContext::get_user_data