Crate bmw_evh

This crate defines and implements the crate::EventHandler. EventHandlers process nonblocking i/o events. They are implemented for linux, windows, and macos. Each platform has a different implementation due to the differences between these platforms. For linux, epoll is used. On macos kqueues are used. On windows, wepoll is used. The result is a cross-platform, performant nonblocking i/o event handler.

Performance

TBD

Using eventhandlers in your project

Add the following to your Cargo.toml:

 bmw_evh = { git = "https://github.com/cgilliard/bitcoinmw"  }

Optionally, you may wish to use the other associated crates:

 bmw_err    = { git = "https://github.com/cgilliard/bitcoinmw"  }
 bmw_log    = { git = "https://github.com/cgilliard/bitcoinmw"  }
 bmw_derive = { git = "https://github.com/cgilliard/bitcoinmw"  }
 bmw_util   = { git = "https://github.com/cgilliard/bitcoinmw"  }

The linux dependencies can be installed with the following commands on ubuntu:

 $ sudo apt-get update -yqq
 $ sudo apt-get install -yqq --no-install-recommends libncursesw5-dev libssl-dev

The macos dependencies can be installed with the following commands

$ brew install llvm

The windows dependencies can be installed with the following commands

$ choco install -y llvm

BitcoinMW is tested with the latest version of rust. Please ensure to update it to the latest version.

Examples

 // Echo Server

 // import the error, log, evh crate and several other things
 use bmw_err::*;
 use bmw_evh::*;
 use bmw_log::*;
 use bmw_test::port::pick_free_port;
 use std::net::TcpStream;
 use std::io::{Read,Write};

 info!();

 fn main() -> Result<(), Error> {
     // create an evh instance with the default configuration
     let mut evh = eventhandler!()?;

     // set the on read handler for this evh
     evh.set_on_read(move |cd, _ctx, _attachment| {
         // log the connection_id of this connection. The connection_id is a random u128
         //value. Each connection has a unique id.
         info!("read data on connection {}", cd.get_connection_id())?;

         // data read is stored in a linked list of slabs. first_slab returns the first
         // slab in the list.
         let first_slab = cd.first_slab();

         // in this example, we don't use it, but we could get the last slab in the list
         // if more than one slab of data may be returned.
         let _last_slab = cd.last_slab();

         // get the slab_offset. This is the offset in the last slab read. The slabs
         // before the last slab will be full so no offset is needed for them. In this
         // example, we always have only a single slab so the offset is always the offset
         // of the slab we are looking at.
         let slab_offset = cd.slab_offset();

         // the borrow slab allocator function allows for the on_read callback to analyze
         // the data that has been read by this connection. The slab_allocator that is
         // passed to the closure is immutable so none of the data can be modified.
         let res = cd.borrow_slab_allocator(move |sa| {
             // get the first slab
             let slab = sa.get(first_slab.try_into()?)?;

             // log the number of bytes that have been read
             info!("read {} bytes", slab_offset)?;

             // create a vec and extend it with the data that was read
             let mut ret: Vec<u8> = vec![];
             ret.extend(&slab.get()[0..slab_offset as usize]);

             // Return the data that was read. The return value is a generic so it
             // could be any type. In this case, we return a Vec<u8>.
             Ok(ret)
         })?;

         // Clear all the data through the first slab, which in this example is assumed
         // to be the last slab. Once this function is called, the subsequent executions
         // of this callback will not include this slab.
         cd.clear_through(first_slab)?;

         // Return a write handle and echo back the data that was read.
         cd.write_handle().write(&res)?;

         Ok(())
     })?;
     evh.set_on_accept(move |cd, _ctx| {
         // The on_accept callback is executed when a connection is accepted.
         info!("accepted connection id = {}", cd.get_connection_id())?;
         Ok(())
     })?;
     evh.set_on_close(move |cd, _ctx| {
         // The on_close callback is executed when a connection is closed.
         info!("closed connection id = {}", cd.get_connection_id())?;
         Ok(())
     })?;
     evh.set_on_panic(move |_ctx, e| {
         // The error is returned by the panic handler as a Box<dyn Any> so we downcast
         // to &str to get the message.
         let e = e.downcast_ref::<&str>().unwrap();
         // The on_panic callback is executed when a thread panic occurs.
         warn!("callback generated thread panic: {}", e)?;
         Ok(())
     })?;
     evh.set_housekeeper(move |_ctx| {
         // The housekeper callback is executed once per thread every second by default.
         info!("Housekeeper executed")?;
         Ok(())
     })?;

     // start the evh
     evh.start()?;

     // pick a free port for our server to bind to
     let (addr, handles) = loop {
         let port = pick_free_port()?;
         info!("using port = {}", port);
         // bind to the loopback interface.
         let addr = format!("127.0.0.1:{}", port).clone();

         // create our server handles for the default 6 threads of the evh.
         // We use a tcp_listener backlog of 10 in this example and we're setting
         // SO_REUSE_PORT to true.
         let handles = create_listeners(6, &addr, 10, true);
         match handles {
             Ok(handles) => break (addr, handles),
             Err(_e) => {}
         }
     };

     // create a ServerConnection with no tls configurations so it will be plain
     // text.
     let sc = ServerConnection {
         tls_config: None,
         handles,
         is_reuse_port: true,
     };

     // add our server connection to the evh.
     evh.add_server(sc, Box::new(""))?;

     // create a client connection to test the evh
     let mut connection = TcpStream::connect(addr)?;

     // send a message "test1".
     connection.write(b"test1")?;

     // assert that the response is an echo of our message.
     let mut buf = vec![];
     buf.resize(100, 0u8);
     let len = connection.read(&mut buf)?;
     assert_eq!(&buf[0..len], b"test1");

     // send a second message "test2".
     connection.write(b"test2")?;

     // assert that the response is an echo of our message.
     let len = connection.read(&mut buf)?;
     assert_eq!(&buf[0..len], b"test2");

     // stop the evh
     evh.stop()?;

     Ok(())
 }

Macros

eventhandler

Build an crate::EventHandler instance. See module level documentation for examples. Optionally, an crate::EventHandlerConfig may be specified. If none is specified, the default values are used.

Structs

AttachmentHolder

Builder

The structure that builds eventhandlers.

ClientConnection

A struct which specifies a client connection.

CloseHandle

A struct which can be used to close a connection (note: if writing is needed as well use WriteHandle, but this is a minimal structure needed to close the connection

ConnectionData

A struct which is passed to several of the callbacks in crate::EventHandler. It provides information on the connection from which data is read.

EventHandlerConfig

The configuration for the crate::EventHandler.

EventHandlerData

ServerConnection

A struct which specifies a server connection.

ThreadContext

A thread context which is passed to the callbacks specified by a crate::EventHandler.

TlsClientConfig

TlsClientConfig specifies the configuration for a tls client.

TlsServerConfig

TlsServerConfig specifies the configuration for a tls server.

WriteHandle

A struct which is used to write to a connection.

WriteState

Constants

READ_SLAB_DATA_SIZE

The size of the data which is stored in read slabs. This data is followed by 4 bytes which is a pointer to the next slab in the list.

Traits

ConnData

This trait which is implemented by crate::ConnectionData. This trait is used to interact with a connection.

EventHandler

This trait defines the behaviour of an eventhandler. See the module level documentation for examples.

Functions

create_listeners

Create listeners for use with the crate::ServerConnection struct. This function crates an array of handles which can be used to construct a crate::ServerConnection object. size is the size of the array. It must be equal to the number of threads that the crate::EventHandler has configured. addr is the socketaddress to bind to. (For example: 127.0.0.1:80 or 0.0.0.0:443.). listen_size is the size of the listener backlog for this tcp/ip connection. reuse_port specifies whether or not to reuse the port on a per thread basis for this connection. This is only available on linux and will be ignored on other platforms.