Function bmw_evh::create_listeners

pub fn create_listeners(
    size: usize,
    addr: &str,
    listen_size: usize,
    reuse_port: bool
) -> Result<Array<RawFd>, Error>

Create listeners for use with the crate::ServerConnection struct. This function crates an array of handles which can be used to construct a crate::ServerConnection object. size is the size of the array. It must be equal to the number of threads that the crate::EventHandler has configured. addr is the socketaddress to bind to. (For example: 127.0.0.1:80 or 0.0.0.0:443.). listen_size is the size of the listener backlog for this tcp/ip connection. reuse_port specifies whether or not to reuse the port on a per thread basis for this connection. This is only available on linux and will be ignored on other platforms.